requirements.txt
asgiref==3.2.3
autopep8==1.4.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.1
django-crispy-forms==1.8.1
django-import-export==2.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==1.4.6
odfpy==1.4.0
openpyxl==3.0.2
Pillow==6.2.1
pycodestyle==2.5.0
pytz==2019.3
PyYAML==5.2
sqlparse==0.3.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0